# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Description de mon API",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "License MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/categorie": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Get all categories",
                "description": "Get all categories",
                "operationId": "2e1b455385734d344b583ff682bf081d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Create a new category",
                "description": "Create a new category",
                "operationId": "28bd6cc293d2213cc109346248df50ee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorie"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categorie/{post}": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Get a single category",
                "description": "Get a single category",
                "operationId": "5197b9f48aed8d5f7d498a912fae9c7e",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Update an existing category",
                "description": "Update an existing category",
                "operationId": "9bc6b53880ba94d61cb48c24e85a7d4b",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorie"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Delete a category",
                "description": "Delete a category",
                "operationId": "ae5ce067f68f761077ed4bf245d4d156",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "description": "Get all products",
                "operationId": "1c3b4d2d08a6f288a591a38ead02e5b4",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "description": "Create a new product",
                "operationId": "11b374caabe03b71870182cdbef8275d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/product/{post}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get a single product",
                "description": "Get a single product",
                "operationId": "98f35bb5fa75208197645c1851fa808e",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update an existing product",
                "description": "Update an existing product",
                "operationId": "6af31266363d854f3a93a3fbffaedab3",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product",
                "description": "Delete a product",
                "operationId": "ce29df30a2a8a5f66b39612bc03c32b7",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/filtreCate/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Filter products by category",
                "description": "Filter products by category ID",
                "operationId": "552e46b97c3ca11a8cb5aff0bb09dcbe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/welcome": {
            "get": {
                "tags": [
                    "Welcome"
                ],
                "summary": "Get all products",
                "description": "Get all products",
                "operationId": "9cca4a2bfb451251214737d05e0f5475",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Categorie": {
                "title": "Categorie",
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "title": {
                        "description": "The title of the category",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the category",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock",
                    "image"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the product",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the product",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "The stock quantity of the product",
                        "type": "integer"
                    },
                    "image": {
                        "description": "The URL of the product image",
                        "type": "string",
                        "format": "uri"
                    }
                },
                "type": "object"
            },
...
        }
    }
}